NVS Storage Validation Loops
The restore configuration validation loops for checking validity of stored information is completed; it re-prompts the user and avoids retrying where it's irrelevant.

Timers and more sophisticated retry methodologies could be implemented, and eventually a page detailing why the re-prompt occurred with the previous data shown on screen will be added as well. But for now the connect to WiFi portion of the firmware is complete.